Changed Id type

+++ Tournaments.Domain/Entities/TournamentTeam.cs
@@ -4,10 +4,10 @@ namespace Tournaments.Domain.Entities
 {
 	public class TournamentTeam
 	{
-        public int TeamId { get; set; }
+        public long TeamId { get; set; }
         public Team Team { get; set; } = null!;
 
-		public int TournamentId { get; set; }
+		public long TournamentId { get; set; }
 		public Tournament Tournament { get; set; } = null!;
 
 		public TournamentTeamStatus Status { get; set; }
